Test/bdd get errors (#2076)
* removed unused dependencies

* convert to esm module, update dependencies

* update test files to mjs

* changes to modules imports and test:bdd command in package.json

* temporary fix for sequelize migrations

* downgrade comunica

* fix lint errors

* management wallets start from the middle of wallets array

* fix multiple issues

* remove logs, add sleep in tests

* fix import issues in unit tests

* add asset type json

* properly import json files

* publishStartError testcase in progress

* run sequelize migrations programmatically

* publishStartError testcase covered

* remove unused dependencies

* remove unused dependencies

* udate dev dependencies

* Added removing of expired cache files

* getAssertionIdError testcase completed. Added selection and try catch blocks to operation-id-service and web3-service

* add dependencies document

* update package-lock.json

* Created new scenario for publishing

* added error logging for fetching latest commit hash

* typos

* update origintrail modules versions

* Updated depnedencies for pokladot api

* Added ganache starting for local network setup

* changed require to imports and created steps-utils module

* changed require to imports and created steps-utils module

* added sleep before GET operation

* added check if file exist in removeFile. Extracting classes instead of functions. Added regex check in step definition

* added new get testcase

* maxRetryCount set to 5

* using json.parse and fs.readFile insted of require for reading .json files.deleted comments

* remove file deletion from initializeConfiguration

+++ src/service/file-service.js
@@ -64,9 +64,13 @@ class FileService {
     }
 
     async removeFile(filePath) {
-        this.logger.debug(`Removing file on path: ${filePath}`);
-        await unlink(filePath);
-        return true;
+        if (await this.fileExists(filePath)) {
+            this.logger.debug(`Removing file on path: ${filePath}`);
+            await unlink(filePath);
+            return true;
+        }
+        this.logger.debug(`File doesn't exist on file path: ${filePath}`);
+        return false;
     }
 
     getDataFolderPath() {
